third_party/WebKit/LayoutTests/fast/css/parsing-object-position-expected.txt

This tests checks that all of the input values for object-position parse correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testComputedStyle(";") is "50% 50%"
PASS testComputedStyle("object-position: 10px;") is "10px 50%"
PASS testComputedStyle("object-position: 10px 10px;") is "10px 10px"
PASS testComputedStyle("object-position: right top;") is "100% 0%"
PASS testComputedStyle("object-position: top right;") is "100% 0%"
PASS test("object-position: inherit;") is "inherit"
PASS test("object-position: initial;") is "initial"
PASS test("object-position: left;") is "0% 50%"
PASS test("object-position: top;") is "50% 0%"
PASS test("object-position: top right;") is "100% 0%"
PASS test("object-position: right top;") is "100% 0%"
PASS test("object-position: center center;") is "50% 50%"
PASS test("object-position: center;") is "50% 50%"
PASS test("object-position: bottom center;") is "50% 100%"
PASS test("object-position: left center;") is "0% 50%"
PASS test("object-position: bottom center;") is "50% 100%"
PASS test("object-position: center left;") is "0% 50%"
PASS test("object-position: center bottom;") is "50% 100%"
PASS test("object-position: 100px;") is "100px 50%"
PASS test("object-position: 100px 100px;") is "100px 100px"
PASS test("object-position: 100px 200px;") is "100px 200px"
PASS test("object-position: -50% 0;") is "-50% 0px"
PASS test("object-position: 3em 0;") is "3em 0px"
PASS test("object-position: left 33px;") is "0% 33px"
PASS test("object-position: center 33px;") is "50% 33px"
PASS test("object-position: 33px center;") is "33px 50%"
PASS test("object-position: 33px bottom;") is "33px 100%"
PASS test("object-position: 1vh 1vw;") is "1vh 1vw"
PASS test("object-position: 100px 100px 100px;") is null
PASS test("object-position: 100px 100px 200px 200px;") is null
PASS test("object-position: top left center;") is null
PASS test("object-position: top top;") is null
PASS test("object-position: top bottom;") is null
PASS test("object-position: 33px left;") is null
PASS test("object-position: top 33px;") is null
PASS test("object-position: inherit inherit;") is null
PASS test("object-position: initial initial;") is null
PASS test("object-position: -webkit-fill-available;") is null
PASS test("object-position: min-content;") is null
PASS test("object-position: intrinsic;") is null
PASS test("object-position: auto;") is null
PASS test("object-position: none;") is null
PASS test("object-position: fill;") is null
PASS successfullyParsed is true

TEST COMPLETE
